+++ src/BackupPolicyInterval.cc
@@ -19,6 +19,7 @@
 #include "Volume.h"
 #include "Device.h"
 #include "Utils.h"
+#include "Errors.h"
 #include "BackupPolicy.h"
 
 /** @brief The @c interval backup policy; backups are separate by a configurable
@@ -28,14 +29,16 @@ public:
   BackupPolicyInterval(): BackupPolicy("interval") {}
 
   void validate(const Volume *volume) const override {
-    parseInteger(get(volume, "min-interval"), 1,
-                 std::numeric_limits<int>::max());
+    if(parseTimeInterval(get(volume, "min-interval"), 1,
+                         std::numeric_limits<int>::max())
+       < 1)
+      throw SyntaxError("min-interval too small");
   }
 
   bool backup(const Volume *volume, const Device *device) const override {
     time_t now = Date::now();
-    int minInterval = parseInteger(get(volume, "min-interval"), 1,
-                                   std::numeric_limits<int>::max());
+    int minInterval = parseTimeInterval(get(volume, "min-interval"), 1,
+                                        std::numeric_limits<int>::max());
     for(const Backup *backup: volume->backups)
       if(backup->rc == 0 && now - backup->time < minInterval
          && backup->deviceName == device->name)